Enabled Spring validation

See merge request galvad/university!24

+++ University/src/main/java/ua/com/foxminded/galvad/university/controllers/ClassroomsController.java
@@ -1,10 +1,10 @@
 package ua.com.foxminded.galvad.university.controllers;
 
-import java.util.List;
-
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,7 +23,7 @@ public class ClassroomsController {
 	private static final String CLASSROOMS_ADD = "classrooms/add";
 	private static final String CLASSROOMS_EDIT = "classrooms/edit";
 	private static final String CLASSROOMS_DELETE = "classrooms/delete";
-	private static final String CLASSROOM = "classroom";
+	private static final String CLASSROOM_DTO = "classroomDTO";
 	private static final String RESULT = "result";
 
 	@Autowired
@@ -33,59 +33,65 @@ public class ClassroomsController {
 
 	@GetMapping()
 	public String findAll(Model model) {
-		List<ClassroomDTO> listOfClassroomDTOs = classroomService.findAll();
-		model.addAttribute("classrooms", listOfClassroomDTOs);
+		model.addAttribute("classrooms", classroomService.findAll());
 		return CLASSROOMS_LIST;
 	}
 
 	@GetMapping("/add")
 	public String create(Model model) {
-		ClassroomDTO classroomDTO = new ClassroomDTO();
-		model.addAttribute("classroomDTO", classroomDTO);
+		model.addAttribute(CLASSROOM_DTO, new ClassroomDTO());
 		return CLASSROOMS_ADD;
 	}
 
 	@PostMapping("/add")
-	public String createDTO(@ModelAttribute("classroomDTO") ClassroomDTO classroomDTO, Model model) {
+	public String createDTO(@Valid ClassroomDTO classroomDTO, BindingResult result, Model model) {
+		if (classroomService.checkIfExists(classroomDTO)) {
+			result.rejectValue("name", "", "The classroom with the same name is already added to the database!");
+		}
+		if (result.hasErrors()) {
+			return CLASSROOMS_ADD;
+		}
 		classroomService.create(classroomDTO);
-		model.addAttribute(CLASSROOM, classroomDTO);
+		model.addAttribute(CLASSROOM_DTO, classroomDTO);
 		model.addAttribute(RESULT, "A classroom was successfully added.");
 		return CLASSROOMS_RESULT;
 	}
 
 	@PostMapping("/edit")
-	public String editDTO(@ModelAttribute("name") String name, Model model) {
-		model.addAttribute("name", name);
+	public String editDTO(String name, Model model) {
+		ClassroomDTO classroomDTO = classroomService.retrieve(name);
+		model.addAttribute(CLASSROOM_DTO, classroomDTO);
+		model.addAttribute("initialName", classroomDTO.getName());
 		return CLASSROOMS_EDIT;
 	}
 
 	@PostMapping("/edit/result")
-	public String editDTOResult(@ModelAttribute("name") String name, @ModelAttribute("initialName") String initialName,
-			Model model) {
-
-		ClassroomDTO initialClassroomDTO = new ClassroomDTO();
-		initialClassroomDTO.setName(initialName);
-		ClassroomDTO updatedClassroomDTO = new ClassroomDTO();
-		updatedClassroomDTO.setName(name);
-
-		classroomService.update(initialClassroomDTO, updatedClassroomDTO);
+	public String editDTOResult(@Valid ClassroomDTO classroomDTO, BindingResult result,
+			@ModelAttribute("initialName") String initialName, Model model) {
+		if (!classroomDTO.getName().equals(initialName) && classroomService.checkIfExists(classroomDTO)) {
+			result.rejectValue("name", "", "The classroom with the same name is already added to the database!");
+		}
+		if (result.hasErrors()) {
+			return CLASSROOMS_EDIT;
+		}
+		ClassroomDTO initialClassroomDTO = classroomService.retrieve(initialName);
+		classroomService.update(initialClassroomDTO, classroomDTO);
 		model.addAttribute(RESULT, "Classroom was successfully updated");
-		model.addAttribute(CLASSROOM, updatedClassroomDTO);
 		return CLASSROOMS_RESULT;
 	}
 
 	@PostMapping("/delete")
-	public String deleteDTO(@ModelAttribute("name") String name, Model model) {
-		model.addAttribute("name", name);
+	public String deleteDTO(String name, Model model) {
+		model.addAttribute(CLASSROOM_DTO, classroomService.retrieve(name));
 		return CLASSROOMS_DELETE;
 	}
 
 	@PostMapping("/delete/result")
-	public String deleteDTOResult(@ModelAttribute("name") String name, Model model) {
-		ClassroomDTO classroomDTO = new ClassroomDTO();
-		classroomDTO.setName(name);
+	public String deleteDTOResult(@Valid ClassroomDTO classroomDTO, BindingResult result, Model model) {
+		if (result.hasErrors()) {
+			return CLASSROOMS_LIST;
+		}
 		classroomService.delete(classroomDTO);
-		model.addAttribute(CLASSROOM, classroomDTO);
 		model.addAttribute(RESULT, "A classroom was successfully deleted.");
 		return CLASSROOMS_RESULT;
 	}
