+++ University/src/main/java/ua/com/foxminded/galvad/university/controllers/LessonsController.java
@@ -2,6 +2,7 @@ package ua.com.foxminded.galvad.university.controllers;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,6 +12,9 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import ua.com.foxminded.galvad.university.dto.ClassroomDTO;
+import ua.com.foxminded.galvad.university.dto.CourseDTO;
+import ua.com.foxminded.galvad.university.dto.GroupDTO;
 import ua.com.foxminded.galvad.university.dto.LessonDTO;
 import ua.com.foxminded.galvad.university.services.ClassroomService;
 import ua.com.foxminded.galvad.university.services.CourseService;
@@ -26,9 +30,14 @@ public class LessonsController {
 	private static final String LESSONS_ADD = "lessons/add";
 	private static final String LESSONS_EDIT = "lessons/edit";
 	private static final String LESSONS_DELETE = "lessons/delete";
-	private static final String LESSON = "lesson";
 	private static final String RESULT = "result";
 	private static final String LESSON_DTO = "lessonDTO";
+	private static final String GROUPS = "groups";
+	private static final String COURSES = "courses";
+	private static final String CLASSROOMS = "classrooms";
+	private static final String START_TIME = "startTime";
+	private static final String DURATION = "duration";
+	private static final String ERROR = "error";
 
 	private final LessonService lessonService;
 	private final GroupService groupService;
@@ -46,49 +55,53 @@ public class LessonsController {
 
 	@GetMapping()
 	public String findAll(Model model) {
-		List<LessonDTO> listOfLessonDTOs = lessonService.findAll();
-		model.addAttribute("lessons", listOfLessonDTOs);
-		model.addAttribute(LESSON_DTO, new LessonDTO());
+		model.addAttribute("lessons", lessonService.findAll());
 		return LESSONS_LIST;
 	}
 
 	@GetMapping("/add")
 	public String create(Model model) {
-		LessonDTO lessonDTO = new LessonDTO();
-		lessonDTO.setStartTime(0);
-		lessonDTO.setDuration(0);
-		model.addAttribute(LESSON_DTO, lessonDTO);
-
-		List<String> listOfGroupNames = new ArrayList<>();
-		listOfGroupNames.add("");
-		groupService.findAll().stream().forEach(s -> listOfGroupNames.add(s.getName()));
-		model.addAttribute("groups", listOfGroupNames);
-
-		List<String> listOfCourseNames = new ArrayList<>();
-		listOfCourseNames.add("");
-		courseService.findAll().stream().forEach(s -> listOfCourseNames.add(s.getName()));
-		model.addAttribute("courses", listOfCourseNames);
-
-		List<String> listOfClassroomNames = new ArrayList<>();
-		listOfClassroomNames.add("");
-		classroomService.findAll().stream().forEach(s -> listOfClassroomNames.add(s.getName()));
-		model.addAttribute("classrooms", listOfClassroomNames);
+		model.addAttribute(GROUPS, groupService.findAll().stream().map(GroupDTO::getName).collect(Collectors.toList()));
+		model.addAttribute(COURSES,
+				courseService.findAll().stream().map(CourseDTO::getName).collect(Collectors.toList()));
+		model.addAttribute(CLASSROOMS,
+				classroomService.findAll().stream().map(ClassroomDTO::getName).collect(Collectors.toList()));
+		model.addAttribute(START_TIME, "01-01-2021 02:00");
+		model.addAttribute(DURATION, "00:30");
 		return LESSONS_ADD;
 	}
 
 	@PostMapping("/add")
-	public String createDTO(@ModelAttribute("group") String groupName, @ModelAttribute("course") String courseName,
-			@ModelAttribute("classroom") String classroomName, @ModelAttribute("startTime") String startTime,
+	public String createDTO(@ModelAttribute("group") String group, @ModelAttribute("course") String course,
+			@ModelAttribute("classroom") String classroom, @ModelAttribute("startTime") String startTime,
 			@ModelAttribute("duration") String duration, Model model) {
-
 		LessonDTO lessonDTO = new LessonDTO();
-		lessonDTO.setCourse(courseService.retrieve(courseName));
-		lessonDTO.setGroup(groupService.retrieve(groupName));
+		lessonDTO.setGroup(groupService.retrieve(group));
+		lessonDTO.setCourse(courseService.retrieve(course));
+		lessonDTO.setClassroom(classroomService.retrieve(classroom));
 		lessonDTO.setStartTime(lessonService.convertDateToMil(startTime));
 		lessonDTO.setDuration(lessonService.convertTimeToMil(duration));
-		lessonDTO.setClassroom(classroomService.retrieve(classroomName));
+		String error = validate(lessonDTO);
+		if (error.isEmpty() && lessonService.checkIfExists(lessonDTO)) {
+			error = "The same lesson is already added to the database!";
+		}
+		if (!error.isEmpty()) {
+			model.addAttribute("group", group);
+			model.addAttribute("course", course);
+			model.addAttribute("classroom", classroom);
+			model.addAttribute(START_TIME, startTime);
+			model.addAttribute(DURATION, duration);
+			model.addAttribute(ERROR, error);
+			model.addAttribute(GROUPS,
+					groupService.findAll().stream().map(GroupDTO::getName).collect(Collectors.toList()));
+			model.addAttribute(COURSES,
+					courseService.findAll().stream().map(CourseDTO::getName).collect(Collectors.toList()));
+			model.addAttribute(CLASSROOMS,
+					classroomService.findAll().stream().map(ClassroomDTO::getName).collect(Collectors.toList()));
+			return LESSONS_ADD;
+		}
 		lessonService.create(lessonDTO);
-		model.addAttribute(LESSON, lessonDTO);
+		model.addAttribute(LESSON_DTO, lessonDTO);
 		model.addAttribute(RESULT, "A lesson was successfully added.");
 		return LESSON_RESULT;
 	}
@@ -99,30 +112,18 @@ public class LessonsController {
 			@ModelAttribute("startTime") String startTimeString, @ModelAttribute("duration") String durationString,
 			Model model) {
 
-		List<String> listOfGroupNames = new ArrayList<>();
-		groupService.findAll().stream().forEach(s -> listOfGroupNames.add(s.getName()));
-		model.addAttribute("groups", listOfGroupNames);
-
-		List<String> listOfCourseNames = new ArrayList<>();
-		courseService.findAll().stream().forEach(s -> listOfCourseNames.add(s.getName()));
-		model.addAttribute("courses", listOfCourseNames);
-
-		List<String> listOfClassroomNames = new ArrayList<>();
-		classroomService.findAll().stream().forEach(s -> listOfClassroomNames.add(s.getName()));
-		model.addAttribute("classrooms", listOfClassroomNames);
-
-		LessonDTO lessonDTO = new LessonDTO();
-		lessonDTO.setStartTime(lessonService.convertDateToMil(startTimeString));
-		lessonDTO.setDuration(lessonService.convertTimeToMil(durationString));
-		model.addAttribute(LESSON_DTO, lessonDTO);
-
+		model.addAttribute(GROUPS, groupService.findAll().stream().map(GroupDTO::getName).collect(Collectors.toList()));
+		model.addAttribute(COURSES,
+				courseService.findAll().stream().map(CourseDTO::getName).collect(Collectors.toList()));
+		model.addAttribute(CLASSROOMS,
+				classroomService.findAll().stream().map(ClassroomDTO::getName).collect(Collectors.toList()));
+		model.addAttribute(START_TIME, startTimeString);
+		model.addAttribute(DURATION, durationString);
 		model.addAttribute("initialGroup", groupName);
 		model.addAttribute("initialCourse", courseName);
 		model.addAttribute("initialClassroom", classroomName);
 		model.addAttribute("initialStartTime", startTimeString);
 		model.addAttribute("initialDuration", durationString);
-		model.addAttribute(LESSON_DTO, lessonDTO);
-
 		return LESSONS_EDIT;
 	}
 
@@ -150,10 +151,36 @@ public class LessonsController {
 		updatedLessonDTO.setCourse(courseService.retrieve(updatedCourseName));
 		updatedLessonDTO.setStartTime(lessonService.convertDateToMil(updatedStartTimeString));
 		updatedLessonDTO.setDuration(lessonService.convertTimeToMil(updatedDurationString));
-
+		String error = "";
+		if (!updatedLessonDTO.equals(initialLessonDTO)) {
+			error = validate(updatedLessonDTO);
+		}
+		if (error.isEmpty() && lessonService.checkIfExists(updatedLessonDTO)) {
+			error = "The same lesson is already added to the database!";
+		}
+		if (!error.isEmpty()) {
+			model.addAttribute(ERROR, error);
+			model.addAttribute(GROUPS,
+					groupService.findAll().stream().map(GroupDTO::getName).collect(Collectors.toList()));
+			model.addAttribute(COURSES,
+					courseService.findAll().stream().map(CourseDTO::getName).collect(Collectors.toList()));
+			model.addAttribute(CLASSROOMS,
+					classroomService.findAll().stream().map(ClassroomDTO::getName).collect(Collectors.toList()));
+			model.addAttribute("group", updatedGroupName);
+			model.addAttribute("course", updatedCourseName);
+			model.addAttribute("classroom", updatedClassroomName);
+			model.addAttribute(START_TIME, updatedStartTimeString);
+			model.addAttribute(DURATION, updatedDurationString);
+			model.addAttribute("initialGroup", initialGroupName);
+			model.addAttribute("initialCourse", initialCourseName);
+			model.addAttribute("initialClassroom", initialClassroomName);
+			model.addAttribute("initialStartTime", initialStartTimeString);
+			model.addAttribute("initialDuration", initialDurationString);
+			return LESSONS_EDIT;
+		}
 		lessonService.update(initialLessonDTO, updatedLessonDTO);
 		model.addAttribute(RESULT, "Lesson was successfully updated");
-		model.addAttribute(LESSON, updatedLessonDTO);
+		model.addAttribute(LESSON_DTO, updatedLessonDTO);
 		return LESSON_RESULT;
 	}
 
@@ -169,18 +196,40 @@ public class LessonsController {
 		lessonDTO.setCourse(courseService.retrieve(courseName));
 		lessonDTO.setStartTime(lessonService.convertDateToMil(startTimeString));
 		lessonDTO.setDuration(lessonService.convertTimeToMil(durationString));
-
-		model.addAttribute(LESSON, lessonDTO);
+		model.addAttribute(LESSON_DTO, lessonDTO);
 		return LESSONS_DELETE;
 	}
 
 	@PostMapping("/delete/result")
 	public String deleteCourseFromDB(@ModelAttribute("lesson") LessonDTO lessonDTO, Model model) {
-
 		lessonDTO.setCourse(courseService.retrieve(lessonDTO.getCourse().getName()));
+		if (!validate(lessonDTO).isEmpty()) {
+			return LESSONS_LIST;
+		}
+
 		lessonService.delete(lessonDTO);
-		model.addAttribute(RESULT, "A lesson was successfully added");
-		model.addAttribute(LESSON, lessonDTO);
+		model.addAttribute(RESULT, "A lesson was successfully deleted.");
+		model.addAttribute(LESSON_DTO, lessonDTO);
 		return LESSON_RESULT;
+	}
+
+	private String validate(LessonDTO lessonDTO) {
+		List<String> listOfErrors = new ArrayList<>();
+		if (lessonDTO.getGroup() == null) {
+			listOfErrors.add("The group field cannot be blank!");
+		}
+		if (lessonDTO.getCourse() == null) {
+			listOfErrors.add("The course field cannot be blank!");
+		}
+		if (lessonDTO.getClassroom() == null) {
+			listOfErrors.add("The classroom field cannot be blank!");
+		}
+		if (lessonDTO.getStartTime() == null || lessonDTO.getStartTime() < 1609459201000l) {
+			listOfErrors.add("The startTime field cannot be blank or earlier than UTC Jan-01-2021 00:00:01!");
+		}
+		if (lessonDTO.getDuration() == null || lessonDTO.getDuration() < 1800000l) {
+			listOfErrors.add("The duration field cannot be blank or less than 30 minutes!");
+		}
+		return listOfErrors.stream().collect(Collectors.joining("<br>", "", ""));
 	}
 }
